Add route for transaction history
Feature: Add route for transaction history

Added a new route '/transaction-history/:address' to the express app. This route allows users to retrieve the transaction history for a specified address. The implementation of the logic to fetch transaction history is yet to be done. This feature enhances the application's functionality by providing users with the ability to track transactions.